Back-End/requirements.txt
Flask>=3.1.0
Flask-SQLAlchemy>=3.1.1
SQLAlchemy>=2.0.35
flask-cors>=4.0.1
python-dotenv>=1.0.1
pyinstaller>=6.11.0
supabase>=2.3.4
